VLM/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic_settings==2.1.0
python-multipart==0.0.6
requests==2.31.0

transformers==4.45.0
accelerate==0.25.0
safetensors==0.4.1
sentencepiece==0.1.99
protobuf==4.25.2

bitsandbytes==0.41.3
Pillow==10.2.0

torch
huggingface_hub
